Move PAM project to bitwarden_license/src/Services per ADR-0032
Relocate the PAM feature to follow the "Break up Core" ADR's Services
layout:

- bitwarden_license/src/Commercial.Pam -> bitwarden_license/src/Services/Pam
- bitwarden_license/test/Commercial.Pam.Test -> bitwarden_license/test/Services/Pam.Test

Adopt the ADR's Service naming: assembly Commercial.Pam -> Pam, root
namespace Bit.Commercial.Pam -> Bit.Services.Pam, and DI extension
AddCommercialPamServices -> AddPamServices.

Update all references: solution file, Api project reference + Startup
wiring, the AddSwaggerGen doc comment, CODEOWNERS, and the consuming
packages.lock.json files (rename only; version pins and RIDs left at
their committed values).
