+++ .github/CODEOWNERS
@@ -110,8 +110,8 @@ src/Admin/Views/Tools @bitwarden/team-billing-dev
 src/Core/Platform/Push/PushType.cs
 
 # PAM
-bitwarden_license/src/Commercial.Pam @bitwarden/team-pam-dev
-bitwarden_license/test/Commercial.Pam.Test @bitwarden/team-pam-dev
+bitwarden_license/src/Services/Pam @bitwarden/team-pam-dev
+bitwarden_license/test/Services/Pam.Test @bitwarden/team-pam-dev
 
 # SDK
 util/RustSdk @bitwarden/team-sdk-sme
