+++ bitwarden-server.slnx
@@ -38,13 +38,13 @@
       <BuildDependency Project="src/Core/Core.csproj" />
     </Project>
     <Project Path="bitwarden_license/src/Commercial.Infrastructure.EntityFramework/Commercial.Infrastructure.EntityFramework.csproj" />
-    <Project Path="bitwarden_license/src/Commercial.Pam/Commercial.Pam.csproj" />
+    <Project Path="bitwarden_license/src/Services/Pam/Pam.csproj" />
     <Project Path="bitwarden_license/src/Scim/Scim.csproj" />
     <Project Path="bitwarden_license/src/Sso/Sso.csproj" />
   </Folder>
   <Folder Name="/test - Bitwarden License/">
     <Project Path="bitwarden_license/test/Commercial.Core.Test/Commercial.Core.Test.csproj" />
-    <Project Path="bitwarden_license/test/Commercial.Pam.Test/Commercial.Pam.Test.csproj" />
+    <Project Path="bitwarden_license/test/Services/Pam.Test/Pam.Test.csproj" />
     <Project Path="bitwarden_license/test/Scim.IntegrationTest/Scim.IntegrationTest.csproj" />
     <Project Path="bitwarden_license/test/Scim.Test/Scim.Test.csproj" />
     <Project Path="bitwarden_license/test/Sso.IntegrationTest/Sso.IntegrationTest.csproj" />
